requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
whitenoise==6.7.0
gunicorn==21.2.0
# Swagger (любой из двух, можно оба)
drf-yasg==1.21.7
drf-spectacular==0.27.2
